requirements_Version3.txt
streamlit
openai
pymupdf
python-docx